POST: create() function working, friend successfully added
